refactor(socket): extract shared GozdSocketClient target as socket client SSOT
intent(socket): SO_NOSIGPIPE 欠落は CLI と test helper が同じ POSIX socket パターンを手写しで重複させ、一方の修正が他方に伝播しない構造から生まれた。重複そのものを除去する
decision(socket): 生 Unix socket NDJSON 送信を新設の軽量 target GozdSocketClient (Darwin + Foundation のみ依存) に集約し、gozd-cli と SocketServerTests が同一実装を呼ぶ。GozdCore に畳まないのは、hook イベントごとに起動される短命 CLI に CPty / Network / FSEvents / git ops をリンクさせない (CLI を軽く保つ設計意図を維持) ため
decision(test): 回帰テスト 2 本を追加。makeClientSocket の SO_NOSIGPIPE 設定を getsockopt で直接ガードする本と、socketpair の peer close 後 write が SocketClientError.write(EPIPE) で throw されることを決定的に再現する本に分離。後者は自前で SO_NOSIGPIPE を張るため、将来 factory から設定が落ちても runner を殺さず前者が clean に fail で捕える
decision(cli): 共有 writeAll が EINTR を retry するため、従来 retry しなかった CLI 側の非対称も解消。connect の ENOENT / ECONNREFUSED → "gozd app is not running" の既存エラー分岐は挙動保存
